% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,chunker-method}
\alias{initialize,chunker-method}
\title{chunker initializer}
\usage{
\S4method{initialize}{chunker}(.Object, path_, sep_, quoted_, has_colnames_,
  has_rownames_, chunksize_, data_format_, columns_classes_)
}
\description{
chunker initializer
}
\keyword{internal}
