% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,reader-method}
\alias{initialize,reader-method}
\title{reader initializer}
\usage{
\S4method{initialize}{reader}(.Object, path_, sep_, has_colnames_,
  has_rownames_, chunksize_, data_format_, columns_classes_)
}
\description{
reader initializer
}
\keyword{internal}
