% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{reader methods}
\alias{reader methods}
\alias{next_chunk}
\alias{next_chunk,reader-method}
\alias{next_chunk,reader-methods}
\alias{get_table}
\alias{get_table,reader-method}
\alias{get_dataframe,}
\alias{reader-methods}
\alias{get_colnames}
\alias{get_colnames,reader-method}
\alias{get_completed}
\alias{get_completed,reader-method}
\alias{get_completed,reader-methods}
\alias{get_type}
\alias{get_type,reader-method}
\alias{get_type,reader-methods}
\title{Manipulation methods for reader objects}
\usage{
next_chunk(obj)

\S4method{next_chunk}{reader}(obj)

get_table(obj)

\S4method{get_table}{reader}(obj)

get_colnames(obj)

\S4method{get_colnames}{reader}(obj)

get_completed(obj)

\S4method{get_completed}{reader}(obj)

get_type(obj)

\S4method{get_type}{reader}(obj)
}
\arguments{
\item{obj}{object of class reader}
}
\description{
Reader objects can be manipulated with the following methods:
\enumerate{
\item{\bold{next_chunk}}{: allows to read the next chunk of a reader object}
\item{\bold{get_table}}{: retrieve the current data chunk contained in the object}
}

In addition, this information can be retrieved from reader objects:
\enumerate{
\item{\bold{get_completed}}{: get the number of rows already read}
\item{\bold{get_colnames}}{: get column names of the reader object}
}
}
\details{
See \code{\link{reader}} for examples.
}
